Decoupled cache eviction from cache settings handler fixes #7958 (#7959)
