+++ src/Orchard.Web/Modules/Orchard.OutputCache/Handlers/CacheSettingsPartHandler.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Globalization;
-using System.Linq;
+﻿using System.Linq;
 using System.Xml.Linq;
 using Orchard.ContentManagement;
 using Orchard.ContentManagement.Handlers;
-using Orchard.Core.Common.Models;
 using Orchard.OutputCache.Models;
 using Orchard.OutputCache.Services;
 
@@ -23,26 +20,8 @@ namespace Orchard.OutputCache.Handlers {
                 part.DefaultCacheGraceTime = 60;
             });
 
-            // Evict cached content when updated, removed or destroyed.
-            OnPublished<IContent>((context, part) => Invalidate(part));
-            OnRemoved<IContent>((context, part) => Invalidate(part));
-            OnDestroyed<IContent>((context, part) => Invalidate(part));
-
             OnExporting<CacheSettingsPart>(ExportRouteSettings);
             OnImporting<CacheSettingsPart>(ImportRouteSettings);
-        }
-
-        private void Invalidate(IContent content) {
-            // Remove any item tagged with this content item ID.
-            _cacheService.RemoveByTag(content.ContentItem.Id.ToString(CultureInfo.InvariantCulture));
-
-            // Search the cache for containers too.
-            var commonPart = content.As<CommonPart>();
-            if (commonPart != null) {
-                if (commonPart.Container != null) {
-                    _cacheService.RemoveByTag(commonPart.Container.Id.ToString(CultureInfo.InvariantCulture));
-                }
-            }
         }
 
         private void ExportRouteSettings(ExportContentContext context, CacheSettingsPart part) {
