+++ src/Orchard.Web/Modules/Orchard.OutputCache/Orchard.OutputCache.csproj
@@ -107,6 +107,7 @@
     <Content Include="Web.config" />
     <Content Include="Scripts\Web.config" />
     <Content Include="Styles\Web.config" />
+    <Compile Include="Handlers\CacheItemInvalidationHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Content Include="Module.txt" />
   </ItemGroup>
